+++ R/fortify-igraph.R
@@ -14,6 +14,10 @@
 #'   good results when the size of the nodes is reasonably small.
 #' @param by a character vector that matches an edge attribute, which will be
 #'   used to generate a data frame that can be plotted with
+#' @param stringsAsFactors whether vertex and edge attributes should be
+#'   converted to factors if they are of class \code{character}. Defaults to
+#'   the value of \code{getOption("stringsAsFactors")}, which is \code{TRUE} by
+#'   default: see \code{\link[base]{data.frame}}.
 #' @param ... additional parameters for the \code{\link[igraph]{layout_}} function
 #'
 #' @export
@@ -23,6 +27,7 @@ fortify.igraph <- function(
   layout = igraph::nicely(),
   arrow.gap = ifelse(igraph::is.directed(model), 0.025, 0),
   by = NULL,
+  stringsAsFactors = getOption("stringsAsFactors"),
   ...
 ) {
   # node placement
@@ -51,7 +56,8 @@ fortify.igraph <- function(
         FUN = igraph::get.vertex.attribute,
         graph = model,
         USE.NAMES = TRUE
-      )
+      ),
+      stringsAsFactors = stringsAsFactors
     )
   }
 
@@ -87,7 +93,8 @@ fortify.igraph <- function(
         FUN = igraph::get.edge.attribute,
         graph = model,
         USE.NAMES = TRUE
-      )
+      ),
+      stringsAsFactors = stringsAsFactors
     )
   }